Update radio buttons
