+++ TF_Settings_Web/src/Components/ControlBar/index.ts
@@ -1,5 +1,6 @@
+import ControlBar from './ControlBar';
 import StatusIndicator from './StatusIndicator/StatusIndicator';
 import TabSelector from './TabSelector/TabSelector';
 import VersionIndicator from './VersionIndicator/VersionIndicator';
 
-export { StatusIndicator, TabSelector, VersionIndicator };
+export { ControlBar, StatusIndicator, TabSelector, VersionIndicator };
