--- TF_Settings_Web/src/Components/Radio/RadioGroup.tsx
@@ -1,42 +1,0 @@
-import interactionClasses from '@/Pages/Interactions/Interactions.module.scss';
-
-import React from 'react';
-
-interface RadioProps {
-    name: string;
-    options: string[];
-    selected: number;
-    onChange: (newValue: string) => void;
-}
-
-export class RadioGroup extends React.Component<RadioProps, {}> {
-    private onChange() {
-        // this function is here purely to pass to the input, preventing it becoming ReadOnly
-    }
-
-    render() {
-        return (
-            <div className={interactionClasses['verticalContainer']}>
-                {this.props.options.map((option, index) => (
-                    <label
-                        key={index}
-                        className={interactionClasses['input-label-container']}
-                        onPointerDown={() => {
-                            this.props.onChange(option);
-                        }}
-                    >
-                        <input
-                            type="radio"
-                            name={this.props.name}
-                            value={option}
-                            onChange={this.onChange}
-                            checked={this.props.selected === index}
-                        />
-                        <span className="checkmark" />
-                        <p> {option} </p>
-                    </label>
-                ))}
-            </div>
-        );
-    }
-}
