+++ TF_Settings_Web/src/Components/index.ts
@@ -1,8 +1,7 @@
 import Alert from './Alert/Alert';
-import ControlBar from './ControlBar/ControlBar';
+import { ControlBar, StatusIndicator, TabSelector, VersionIndicator } from './ControlBar';
 import DocsLink from './DocsLink/DocsLink';
-import { RadioGroup } from './Radio/RadioGroup';
-import { RadioLine } from './Radio/RadioLine';
+import { RadioLine, RadioGroup } from './Radio/RadioButtons';
 import { Slider } from './Slider/Slider';
 import { TextSlider } from './Slider/TextSlider';
 import {
@@ -30,6 +29,9 @@ export {
     HorizontalIconTextButton,
     ToggleSwitch,
     ControlBar,
+    StatusIndicator,
+    TabSelector,
+    VersionIndicator,
     LabelledToggleSwitch,
     BaseTFButton,
     DocsLink,
